+++ auto-service-plugin/src/test/resources/fixtures/agp8/app/src/main/java/test/sample/ServiceImpl.java
@@ -1,0 +1,10 @@
+package test.sample;
+
+import com.anymore.auto.AutoService;
+
+@AutoService(Runnable.class)
+public final class ServiceImpl implements Runnable {
+    @Override
+    public void run() {
+    }
+}